Time: 0 ms (100.00%) | Memory: 8 MB (19.97%) - LeetSync

+++ 231-power-of-two/power-of-two.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPowerOfTwo(int n) {
-        long long m=n;
+        long m=n;
         if(m==0) return false;
         if((m&(m-1))==0) return true;
         else return false;
